+++ beacon/blockchain/types.go
@@ -248,8 +248,8 @@ type TelemetrySink interface {
 	MeasureSince(key string, start time.Time, args ...string)
 }
 
-//nolint:revive // We will remove this once generics are cleaned up
-type BlockchainI[GenesisT any] interface {
+//nolint:revive // its ok
+type BlockchainI interface {
 	ProcessGenesisData(
 		context.Context, []byte) (transition.ValidatorUpdates, error)
 	ProcessProposal(
